% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{sm_add_coords_to_length}
\alias{sm_add_coords_to_length}
\title{Add coordinates to length data}
\usage{
sm_add_coords_to_length(res)
}
\arguments{
\item{res}{A list}
}
\value{
A list
}
\description{
Add coordinates to length data
}
