% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sm_bubble_plot}
\alias{sm_bubble_plot}
\title{Plot abundance or biomass}
\usage{
sm_bubble_plot(d)
}
\arguments{
\item{d}{A tibble containing lon, lat, abundance an biomass}
}
\value{
a plot
}
\description{
Plot abundance or biomass
}
