% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_munge.R
\name{sm_import_setup}
\alias{sm_import_setup}
\title{Import setup files: stillingar and stodtoflur}
\usage{
sm_import_setup(res, stillingar, stodtoflur)
}
\arguments{
\item{res}{List returned by \code{sm_import_current()}.}

\item{stillingar}{Path to the stillingar zip file.}

\item{stodtoflur}{Path to the stodtoflur zip file.}
}
\value{
\code{res} extended with \code{res$stillingar} and \code{res$stodtoflur}.
}
\description{
Reads the "stillingar" zip file, renames columns in \code{sti_rallstodvar},
converts coordinates, and reads the "stodtoflur" zip file.
}
\keyword{internal}
