% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_plot.R
\name{sm_length_plot}
\alias{sm_length_plot}
\title{Length frequency plot}
\usage{
sm_length_plot(d, ylab = "xx")
}
\arguments{
\item{d}{A length dataframe}

\item{ylab}{Y-axis label}
}
\value{
a ggplot
}
\description{
Length frequency plot
}
