% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sm_lgw_plot}
\alias{sm_lgw_plot}
\title{Length gutted weight qc-plot}
\usage{
sm_lgw_plot(res, species = 1)
}
\arguments{
\item{res}{list of munged hafvog's tables}

\item{species}{Numerical, species to plot}
}
\value{
A plot
}
\description{
Length gutted weight qc-plot
}
