% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_munge.R
\name{sm_munge}
\alias{sm_munge}
\title{Prepares data for the smxapp}
\usage{
sm_munge(
  maelingar,
  stillingar,
  current.year = lubridate::year(lubridate::today())
)
}
\arguments{
\item{maelingar}{Names of hafvog zip files to be imported}

\item{stillingar}{Name of "stillingar" zip file to be imported}

\item{current.year}{The current survey year}
}
\value{
a list
}
\description{
Prepares data for the smxapp
}
