% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sm_plot_bubble}
\alias{sm_plot_bubble}
\title{Plot abundance or biomass}
\usage{
sm_plot_bubble(d)
}
\arguments{
\item{d}{A tibble containing lon, lat, abundance an biomass}
}
\value{
a plot
}
\description{
Plot abundance or biomass
}
