% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sm_plot_glyph}
\alias{sm_plot_glyph}
\title{Species indices by rectangle}
\usage{
sm_plot_glyph(d)
}
\arguments{
\item{d}{A tibble with data by rectangle}
}
\value{
A plot
}
\description{
Species indices by rectangle
}
