% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sm_plot_length_ungutted}
\alias{sm_plot_length_ungutted}
\title{Length ungutted weight qc-plot}
\usage{
sm_plot_length_ungutted(res, species = 1)
}
\arguments{
\item{res}{list of munged hafvog's tables}

\item{species}{Numerical, species to plot}
}
\value{
A plot
}
\description{
Length ungutted weight qc-plot
}
