% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sm_plot_probability}
\alias{sm_plot_probability}
\title{Species probability of capture}
\usage{
sm_plot_probability(d)
}
\arguments{
\item{d}{A tibble}
}
\value{
A plot
}
\description{
Species probability of capture
}
