% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{sm_plot_weights}
\alias{sm_plot_weights}
\title{Ratio of gutted to ungutted as a function of length}
\usage{
sm_plot_weights(res, species = 1)
}
\arguments{
\item{res}{list of munged hafvog's tables}

\item{species}{Numerical, species to plot}
}
\value{
A plot
}
\description{
Ratio of gutted to ungutted as a function of length
}
