% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sm_qc_kvarnir}
\alias{sm_qc_kvarnir}
\title{Create a table where measurements are susspect}
\usage{
sm_qc_kvarnir(res)
}
\arguments{
\item{res}{A list}
}
\description{
Create a table where measurements are susspect
}
