% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{sm_standardize_by_tow}
\alias{sm_standardize_by_tow}
\title{Standardize by towlength}
\usage{
sm_standardize_by_tow(res, tow = c(2, 4, 8))
}
\arguments{
\item{res}{The hafvog list}

\item{tow}{A character vector specifying min, standard and max towlength}
}
\value{
a list
}
\description{
Standardize by towlength
}
