% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{sm_table_kvarnir}
\alias{sm_table_kvarnir}
\title{Create a table where measurements are susspect}
\usage{
sm_table_kvarnir(res)
}
\arguments{
\item{res}{A list}
}
\value{
a DT table
}
\description{
Create a table where measurements are susspect
}
